Implement navigation and introduction screens
This update introduces three introduction screens for new users, as well as a revamped navigation system. The app will now first check whether the user is new by inspecting the 'user' entry in AsyncStorage. Depending on the outcome, the user is directed either to the new user introduction flow or the home screen. The introduction screens provide information on the app's functionality and usage.